Fix error in mpv command line options
The syntax used in the mpv command line is no longer valid, so I updated
it to use the current syntax.

The code that populates the `{file}` and `{offset}` parameters in the
command assumed those would appear as separate tokens, but that's no
longer the case with the new syntax. Therefore, I also fixed that code
to handle the more general case where the parameter somewhere within a
token.